# Conflicts:
#	app/build.gradle
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/jkabe/app/box/adapter/AsetsAdapter.java
#	app/src/main/java/com/jkabe/app/box/adapter/AssetsAdapter.java
#	app/src/main/java/com/jkabe/app/box/adapter/IncomeAdapter.java
#	app/src/main/java/com/jkabe/app/box/adapter/TraverAdapter.java
#	app/src/main/java/com/jkabe/app/box/base/BaseApplication.java
#	app/src/main/java/com/jkabe/app/box/bean/UserInfo.java
#	app/src/main/java/com/jkabe/app/box/box/AssetsActivity.java
#	app/src/main/java/com/jkabe/app/box/box/DrawActivity.java
#	app/src/main/java/com/jkabe/app/box/box/IncomeActivity.java
#	app/src/main/java/com/jkabe/app/box/box/RechargeActivity.java
#	app/src/main/java/com/jkabe/app/box/config/Api.java
#	app/src/main/java/com/jkabe/app/box/config/okHttpModel.java
#	app/src/main/java/com/jkabe/app/box/ui/ActivationActivity.java
#	app/src/main/java/com/jkabe/app/box/ui/BindActivity.java
#	app/src/main/java/com/jkabe/app/box/ui/BrandCarActivity.java
#	app/src/main/java/com/jkabe/app/box/ui/InvitationActivity.java
#	app/src/main/java/com/jkabe/app/box/ui/LocationIndexActivity.java
#	app/src/main/java/com/jkabe/app/box/ui/LoginActivity.java
#	app/src/main/java/com/jkabe/app/box/ui/MainActivity.java
#	app/src/main/java/com/jkabe/app/box/ui/RegisterActivity.java
#	app/src/main/java/com/jkabe/app/box/ui/UserActivity.java
#	app/src/main/java/com/jkabe/app/box/ui/fragment/AssetsFragmnt.java
#	app/src/main/java/com/jkabe/app/box/ui/fragment/CarFragment.java
#	app/src/main/java/com/jkabe/app/box/ui/fragment/MeFragment.java
#	app/src/main/java/com/jkabe/app/box/ui/fragment/MineFragment.java
#	app/src/main/java/com/jkabe/app/box/ui/fragment/MiningFragmnt.java
#	app/src/main/java/com/jkabe/app/box/util/Constants.java
#	app/src/main/java/com/jkabe/app/box/util/FileManager.java
#	app/src/main/java/com/jkabe/app/box/util/FileUtil.java
#	app/src/main/java/com/jkabe/app/box/util/JsonParse.java
#	app/src/main/java/com/jkabe/app/box/util/SaveUtils.java
#	app/src/main/java/com/jkabe/app/box/util/SystemTools.java
#	app/src/main/java/com/jkabe/app/box/weight/DialogUtils.java
#	app/src/main/java/com/jkabe/app/box/weight/UpdateManager.java
#	app/src/main/res/layout/activity_assets.xml
#	app/src/main/res/layout/activity_bind.xml
#	app/src/main/res/layout/activity_draw.xml
#	app/src/main/res/layout/activity_income.xml
#	app/src/main/res/layout/activity_location_map.xml
#	app/src/main/res/layout/activity_login.xml
#	app/src/main/res/layout/activity_recharge.xml
#	app/src/main/res/layout/custom_info_window.xml
#	app/src/main/res/layout/fragment_assets.xml
#	app/src/main/res/layout/fragment_me.xml
#	app/src/main/res/layout/fragment_mining.xml
#	app/src/main/res/layout/item_income.xml
#	app/src/main/res/mipmap-xhdpi/icon_car.png
#	app/src/main/res/mipmap-xxhdpi/icon_car.png
#	app/src/main/res/values/strings.xml
#	app/src/main/res/values/styles.xml
#	build.gradle
#	library1/build/intermediates/compile_only_not_namespaced_r_class_jar/debug/R.jar
#	library1/build/intermediates/compile_only_not_namespaced_r_class_jar/release/R.jar
#	library1/build/intermediates/incremental/packageDebugResources/compile-file-map.properties
#	library1/build/intermediates/incremental/packageReleaseResources/compile-file-map.properties
#	library1/build/intermediates/manifest_merge_blame_file/debugAndroidTest/manifest-merger-blame-debug-androidTest-report.txt

+++ app/src/main/java/com/jkabe/app/box/glide/GlideRoundTransform.java
@@ -11,7 +11,6 @@ import android.graphics.RectF;
 import com.bumptech.glide.load.engine.bitmap_recycle.BitmapPool;
 import com.bumptech.glide.load.resource.bitmap.BitmapTransformation;
 import com.bumptech.glide.load.resource.bitmap.TransformationUtils;
-import com.jkabe.app.box.util.LogUtils;
 
 import java.security.MessageDigest;
 
@@ -23,7 +22,7 @@ public class GlideRoundTransform extends BitmapTransformation {
 
     public static float radius = 0f;
 
-    public GlideRoundTransform(Context context, float dp) {
+    public GlideRoundTransform(Context context, int dp) {
         super(context);
         radius = Resources.getSystem().getDisplayMetrics().density * dp;
     }
